Move clobber of test root into a stand alone function
Dave L. found the UI for clobbering the test root to be confusing
because it was asking ever time a testspec was deleted, but we didn't
want to remove it every time. Instead we keep a list of test roots
that are eligible to be clobbered, then remove them at the end after
all the test cases and data have already been removed.

Test suite: clobbered a bunch of test directories